Added Purple Support
Added reference to a custom purple style. The themes lacked a true deep purple color, so I've quickly added one in based on my previous edits for Air legacy... Tested and working without any issues. Optionally can be called "darkpurple" as it is darker than actual purple (of which there are many variations).